refactor: FakeFixtures 분리
